Change fields/ to preview_panel/fields/ to better fit with the file attributes refactor
