+++ src/tagstudio/qt/views/preview_panel/fields/field_container.py
@@ -10,21 +10,21 @@ from PySide6.QtGui import QEnterEvent, QPixmap, QResizeEvent, Qt
 from PySide6.QtWidgets import QHBoxLayout, QLabel, QPushButton, QVBoxLayout, QWidget
 
 from tagstudio.core.enums import Theme
-from tagstudio.qt.views.fields.field_widget import FieldWidget
+from tagstudio.qt.views.preview_panel.fields.field_widget import FieldWidget
 
 # TODO: reference a resources folder rather than path.parents[2]?
 clipboard_icon_128: Image.Image = Image.open(
-    str(Path(__file__).parents[3] / "resources/qt/images/clipboard_icon_128.png")
+    str(Path(__file__).parents[4] / "resources/qt/images/clipboard_icon_128.png")
 ).resize((math.floor(24 * 1.25), math.floor(24 * 1.25)))
 clipboard_icon_128.load()
 
 edit_icon_128: Image.Image = Image.open(
-    str(Path(__file__).parents[3] / "resources/qt/images/edit_icon_128.png")
+    str(Path(__file__).parents[4] / "resources/qt/images/edit_icon_128.png")
 ).resize((math.floor(24 * 1.25), math.floor(24 * 1.25)))
 edit_icon_128.load()
 
 trash_icon_128: Image.Image = Image.open(
-    str(Path(__file__).parents[3] / "resources/qt/images/trash_icon_128.png")
+    str(Path(__file__).parents[4] / "resources/qt/images/trash_icon_128.png")
 ).resize((math.floor(24 * 1.25), math.floor(24 * 1.25)))
 trash_icon_128.load()
