+++ src/tagstudio/qt/views/preview_panel/fields/field_widget.py
